Support for Top K Top P sampling kernel (#369) (#230)
* Support for Top K Top P sampling kernel (#369)

* added support for top_p_renorm_probs kernel

* added support for top_k_top_sampling

* fix lint error

* fix lint error

* added support for top_p_arr test case

* chnage top_k dtype to int

* change the schema , making output as a first arg

* added support for bf16 and float16 dtype for top_p_renorm

* optimize joint top-k/top-p sampling and align with min-p helpers
Reuse the shared sgl::sampling helpers (comm/Sampling.h)

* simplify top_p_renorm_probs to fp32-only

* use sycle native recips, set pragma unroll to 2, flatten the branching and use sycl::select

* isolate the topk topp sampling probs kernel

* lint fix

* Add fix top_k/top_p per-request array indexing in joint sampling kernel and proper checks

* Add per-request array test for joint top_k/top_p sampling
Extract a shared torch_top_k_top_p_joint_mask ground-truth helper
Also generalize TopKTopPSamplingKernel's work-group size
switch its bool dispatch to AT_DISPATCH_BOOL_NO_RETURN

* Add missing XPU symbol

* lint fix

---------


* remove the unwanted symbol bug

---------

+++ include/sgl_kernel_ops.h
@@ -506,7 +506,7 @@ void top_k_top_p_sampling_from_probs(
     at::Tensor output,
     std::optional<at::Tensor> maybe_indices,
     std::optional<at::Tensor> maybe_top_k_arr,
-    double top_k_val,
+    int64_t top_k_val,
     std::optional<at::Tensor> maybe_top_p_arr,
     double top_p_val,
     bool deterministic,
